pyproject.toml
[tool.poetry]
name = "aws-ops-networking"
version = "0.1.0"
description = "Foundation VPC networking stack for ops-lab platform"
authors = ["simoda"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
aws-cdk-lib = "^2.180.0"
constructs = "^10.0.0"
boto3 = "^1.38.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
